Debugging

+++ src/net/kevxu/senselib/StepDetector.java
@@ -83,12 +83,12 @@ public class StepDetector implements SensorEventListener {
 		}
 
 		public float getAccelInGravityDirection(float[] linearAccel, float[] gravity) {
-			double gravityScalar = Math.sqrt(gravity[0] * gravity[0]
+			float gravityScalar = (float) Math.sqrt(gravity[0] * gravity[0]
 					+ gravity[1] * gravity[1] + gravity[2] * gravity[2]);
 			float dotProduct = linearAccel[0] * gravity[0] + linearAccel[1]
 					* gravity[1] + linearAccel[2] * gravity[2];
 
-			return (float) (dotProduct / gravityScalar);
+			return dotProduct / gravityScalar;
 		}
 
 		@Override
